Add audience to token generator
Certain implementations of oauth2, such as auth0, require the audience
parameter to be send to the token generation endpoint.

This commit adds an additional parameter to, by default, send the
audience parameter.

This should be be broadly compatible and ignored by other
implementations. Otherwise, we can consider an additional configuration
option