Retry retrieving an access token with exponential backoff
When the OIDC provider is unavailable when a new token is requested, the client
is stuck for an indefinite state. This is only resolved by restarting the client or
the server.

With this feature, we want to achieve a scenario where a token is requested for a
set amount of retries when the OIDC provider is not available.